Last Sync: 2025-08-17 18:53 (Mobile)

+++ src/plugins/remark-link-card.ts
@@ -329,7 +329,7 @@ const downloadImage = async (
     const filePath = path.join(cachePath, fileName)
     const buffer = Buffer.from(arrayBuffer)
 
-    await fs.writeFile(filePath, buffer)
+    await fs.writeFile(filePath, buffer as Uint8Array)
 
     return fileName
   } catch (error) {
